requirements.txt
flask_restful
requests
flask
psycopg2-binary
pandas
flask
gunicorn
